% Generated by roxygen2 (4.0.1): do not edit by hand
\name{delete}
\alias{delete}
\title{delete}
\usage{
output = delete(object,id)
}
\arguments{
\item{object}{A Vtiger object}

\item{id}{A Vtiger id (of type character) of the element to be
retrieved}
}
\value{
A list describing whether or not the delete was successful
}
\description{
Delete an entry from the Vtiger system.
}
\examples{
\dontrun{
output = delete(object,id="12x3") ## Make sure this is a valid id
output$status ## Should be = "successful"}
}

