% Generated by roxygen2 (4.0.1): do not edit by hand
\name{describe}
\alias{describe}
\title{describe}
\usage{
output = describe(object,"Contacts")
output = describe(object,"Accounts")
}
\arguments{
\item{object}{A Vtiger object}

\item{elementType}{A string for valid Vtiger component}
}
\value{
A list containing information on all fields contained within the
Vtiger object given by elementType, together with access permissions
(nullable, editable, mandatory, etc).
}
\description{
Get the type information about a given Vtiger object
}

