% Generated by roxygen2 (4.0.1): do not edit by hand
\name{listtypes}
\alias{listtypes}
\title{listtypes}
\usage{
output = listtypes(object)
}
\arguments{
\item{object}{A Vtiger object}
}
\value{
A list containing two components. The first component ("types")
contains a character vector of names of Vtiger objects. The second component
("information") is a list where each element contains information on each on
Vtiger objects contained in "types"
}
\description{
List the names of all the Vtiger objects available through the
api.
}

