% Generated by roxygen2 (4.0.1): do not edit by hand
\name{login}
\alias{login}
\title{login}
\usage{
object = login(object)
}
\arguments{
\item{object}{A Vtiger object}
}
\value{
Returns Vtiger object with valid, populated sessionName slot
}
\description{
Login to Vtiger CRM system. userName, accessKey and
webServiceUrl slots must be populated with valid values.
}

