% Generated by roxygen2 (4.0.1): do not edit by hand
\name{logout}
\alias{logout}
\title{logout}
\usage{
output = logout(object)
}
\arguments{
\item{object}{A Vtiger object}
}
\value{
A logical indicating whether logout was successful.
}
\description{
Logout from the webservices session. This leaves the webservice
session invalid for further use.
}
\examples{
\dontrun{
output = logout(object) ## output should be = TRUE}
}

