% Generated by roxygen2 (4.0.1): do not edit by hand
\name{query}
\alias{query}
\title{query}
\usage{
output = query(object,q)
}
\arguments{
\item{object}{A Vtiger object}

\item{q}{A character string containing a valid Vtiger (SQL-like) query}
}
\value{
A list containing all elements selected by the query
}
\description{
The query operation provides a way to query the Vtiger system
for data.
}
\examples{
\dontrun{
output = query(object,q="SELECT * FROM Contacts;")}
}

