% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sync}
\alias{sync}
\title{sync}
\usage{
output = sync(object,modifiedTime,elementType)
}
\arguments{
\item{object}{A Vtiger object}

\item{modifiedTime}{A UNIX time (integer, number of seconds since 1970).
Inputs of other mode will be coerced to UNIX time.}
}
\value{
A list containing all elements from the Vtiger object
\code{elementType} that have been updated since \code{modifiedTime}.
}
\description{
Sync return a list containing details of all changes in Vtiger
system after modifiedTime.
}
\examples{
\dontrun{
output = sync(object,as.numeric(as.POSIXct("2014-07-05 00:00")))}
}

